+++ src/plugins/data_views/public/plugin.ts
@@ -28,7 +28,7 @@ import { getIndices, HasData } from './services';
 import { debounceByKey } from './debounce_by_key';
 
 import { DATA_VIEW_SAVED_OBJECT_TYPE } from '../common/constants';
-import { LATEST_VERSION } from '../common/content_management/v1/constants';
+import { LATEST_VERSION } from '../common/content_management/constants';
 
 export class DataViewsPublicPlugin
   implements
