+++ src/plugins/data_views/server/content_management/schemas/v1/cm_services.ts
@@ -16,9 +16,9 @@ import {
   createResultSchema,
   searchOptionsSchemas,
 } from '@kbn/content-management-utils';
-import { DataViewType } from '../..';
-import { MAX_DATA_VIEW_FIELD_DESCRIPTION_LENGTH } from '../../constants';
-import { serializedFieldFormatSchema, fieldSpecSchema } from '../../schemas';
+import { DataViewType } from '../../../../common';
+import { MAX_DATA_VIEW_FIELD_DESCRIPTION_LENGTH } from '../../../../common/constants';
+import { serializedFieldFormatSchema, fieldSpecSchema } from '../../../rest_api_routes/schemas';
 
 const dataViewAttributesSchema = schema.object(
   {
