+++ src/plugins/data_views/server/content_management/schemas/v1/index.ts
