+++ src/plugins/data_views/server/rest_api_routes/public/runtime_fields/update_runtime_field.ts
@@ -13,7 +13,7 @@ import { DataViewsService } from '../../../../common/data_views';
 import { RuntimeField } from '../../../../common/types';
 import { ErrorIndexPatternFieldNotFound } from '../../../error';
 import { handleErrors } from '../util/handle_errors';
-import { runtimeFieldSchemaUpdate } from '../../../../common/schemas';
+import { runtimeFieldSchemaUpdate } from '../../schemas';
 import type {
   DataViewsServerPluginStart,
   DataViewsServerPluginStartDependencies,
