+++ src/plugins/data_views/server/rest_api_routes/public/scripted_fields/update_scripted_field.ts
@@ -11,7 +11,7 @@ import { IRouter, StartServicesAccessor } from '@kbn/core/server';
 import { FieldSpec } from '../../../../common';
 import { ErrorIndexPatternFieldNotFound } from '../../../error';
 import { handleErrors } from '../util/handle_errors';
-import { fieldSpecSchemaFields } from '../../../../common/schemas';
+import { fieldSpecSchemaFields } from '../../schemas';
 import type {
   DataViewsServerPluginStart,
   DataViewsServerPluginStartDependencies,
